refactor(client): extract reusable parts of batch request handling
Move parts of the batch request handling logic that can be reused in
tests in `prisma-engines` to `@prisma/client-engine-runtime`.

Ref: https://github.com/prisma/prisma-engines/pull/5590
Ref: https://linear.app/prisma-company/issue/ORM-1225/implement-batching-support-in-qc-tests-in-prisma-engines
