+++ packages/client-common/src/QueryCompiler.ts
@@ -1,20 +1,5 @@
+import type { BatchResponse } from '@prisma/client-engine-runtime'
 import { ConnectionInfo, Provider } from '@prisma/driver-adapter-utils'
-
-export type BatchResponse = MultiBatchResponse | CompactedBatchResponse
-
-export type MultiBatchResponse = {
-  type: 'multi'
-  plans: object[]
-}
-
-export type CompactedBatchResponse = {
-  type: 'compacted'
-  plan: {}
-  arguments: Record<string, {}>[]
-  nestedSelection: string[]
-  keys: string[]
-  expectNonEmpty: boolean
-}
 
 export type QueryCompiler = {
   compile(request: string): {}
