+++ packages/client/src/runtime/core/types/exported/JsApi.ts
@@ -1,10 +1,11 @@
 import type * as DMMF from '@prisma/dmmf'
-import Decimal from 'decimal.js'
 
 import { DecimalJsLike } from './DecimalJsLike'
 import { FieldRef } from './FieldRef'
 import { ObjectEnumValue } from './ObjectEnums'
 import { Skip } from './Skip'
+
+export type { JsOutputValue } from '@prisma/client-engine-runtime'
 
 export type Action = keyof typeof DMMF.ModelAction | 'executeRaw' | 'queryRaw' | 'runCommandRaw'
 
@@ -45,15 +46,3 @@ export type RawParameters = {
   __prismaRawParameters__: true
   values: string
 }
-
-export type JsOutputValue =
-  | null
-  | string
-  | number
-  | boolean
-  | bigint
-  | Uint8Array
-  | Date
-  | Decimal
-  | JsOutputValue[]
-  | { [key: string]: JsOutputValue }
